🚚 refator: utils 폴더 생성 후, 시간 함수 export
